+++ gulpfile.js
@@ -15,6 +15,25 @@ const dartSass = require('sass-embedded')
 const config = require('./app/config')
 const { findAvailablePort } = require('./lib/utils')
 
+/**
+ * Add environment variables from .env file
+ * (Requires Node.js v20.12.0 or later)
+ *
+ * @see {@link https://nodejs.org/api/process.html#processloadenvfilepath}
+ */
+if ('loadEnvFile' in process) {
+  try {
+    process.loadEnvFile()
+  } catch (error) {
+    if (error.code === 'ENOENT') {
+      // File not found - this is fine
+    } else {
+      // Some other error occurred
+      throw error
+    }
+  }
+}
+
 // Set configuration variables
 const port = parseInt(process.env.PORT || config.port, 10) || 2000